# storage/command_queue: use a channel-based API