storage/command_queue: use a channel-based API

This permits us to compose waiting for the command queue with a
context cancellation channel, which is also done here.

By lazily creating channels, performance remains unchanged.
```
name                            old time/op    new time/op    delta
InsertDistinct1_Cockroach-24      1.61ms ± 1%    1.64ms ± 4%  +1.79%        (p=0.007 n=10+10)
InsertDistinct10_Cockroach-24     1.14ms ± 2%    1.17ms ± 3%  +2.54%        (p=0.000 n=10+10)
InsertDistinct100_Cockroach-24    1.47ms ± 1%    1.51ms ± 3%  +2.53%         (p=0.003 n=8+10)
KVInsert1_Native-24                663µs ± 1%     664µs ± 2%    ~           (p=1.000 n=10+10)
KVInsert1_SQL-24                   502µs ± 3%     503µs ± 1%    ~            (p=0.360 n=10+8)
KVInsert10_Native-24               820µs ± 1%     820µs ± 1%    ~           (p=0.853 n=10+10)
KVInsert10_SQL-24                  687µs ± 2%     686µs ± 2%    ~           (p=0.971 n=10+10)
KVInsert100_Native-24             1.97ms ± 1%    1.94ms ± 1%  -1.22%         (p=0.004 n=10+8)
KVInsert100_SQL-24                2.02ms ± 2%    2.01ms ± 1%    ~           (p=0.190 n=10+10)
KVInsert1000_Native-24            11.8ms ± 4%    11.5ms ± 3%    ~           (p=0.063 n=10+10)
KVInsert1000_SQL-24               14.3ms ± 5%    14.0ms ± 3%    ~           (p=0.063 n=10+10)
KVInsert10000_Native-24            119ms ± 5%     120ms ± 5%    ~            (p=0.549 n=9+10)
KVInsert10000_SQL-24               147ms ± 6%     149ms ± 3%    ~           (p=0.280 n=10+10)

name                            old alloc/op   new alloc/op   delta
InsertDistinct1_Cockroach-24       304kB ± 0%     303kB ± 0%  -0.33%        (p=0.000 n=10+10)
InsertDistinct10_Cockroach-24      316kB ± 0%     318kB ± 2%    ~            (p=0.173 n=8+10)
InsertDistinct100_Cockroach-24     474kB ± 0%     472kB ± 0%  -0.33%         (p=0.000 n=10+9)
KVInsert1_Native-24               37.1kB ± 0%    37.0kB ± 0%  -0.22%         (p=0.000 n=10+9)
KVInsert1_SQL-24                  30.7kB ± 0%    30.6kB ± 0%  -0.24%         (p=0.000 n=10+9)
KVInsert10_Native-24              93.2kB ± 0%    93.0kB ± 0%  -0.28%        (p=0.000 n=10+10)
KVInsert10_SQL-24                 79.3kB ± 0%    79.1kB ± 0%  -0.33%        (p=0.000 n=10+10)
KVInsert100_Native-24              607kB ± 0%     605kB ± 0%  -0.29%         (p=0.000 n=9+10)
KVInsert100_SQL-24                 527kB ± 0%     525kB ± 0%  -0.35%         (p=0.000 n=10+9)
KVInsert1000_Native-24            5.14MB ± 0%    5.13MB ± 0%  -0.31%         (p=0.000 n=10+7)
KVInsert1000_SQL-24               4.58MB ± 0%    4.57MB ± 0%  -0.35%         (p=0.000 n=9+10)
KVInsert10000_Native-24           79.1MB ± 0%    79.0MB ± 0%  -0.20%        (p=0.000 n=10+10)
KVInsert10000_SQL-24              66.3MB ± 0%    66.1MB ± 0%  -0.26%         (p=0.000 n=9+10)

name                            old allocs/op  new allocs/op  delta
InsertDistinct1_Cockroach-24       1.91k ± 0%     1.91k ± 0%    ~            (p=0.507 n=9+10)
InsertDistinct10_Cockroach-24      1.98k ± 0%     2.00k ± 3%    ~            (p=0.214 n=8+10)
InsertDistinct100_Cockroach-24     3.32k ± 0%     3.32k ± 0%  -0.07%         (p=0.047 n=10+9)
KVInsert1_Native-24                  364 ± 0%       364 ± 0%    ~           (p=0.087 n=10+10)
KVInsert1_SQL-24                     314 ± 0%       314 ± 0%    ~     (all samples are equal)
KVInsert10_Native-24                 596 ± 0%       596 ± 0%    ~     (all samples are equal)
KVInsert10_SQL-24                    586 ± 0%       586 ± 0%    ~     (all samples are equal)
KVInsert100_Native-24              2.71k ± 0%     2.71k ± 0%  -0.01%         (p=0.029 n=10+7)
KVInsert100_SQL-24                 3.05k ± 0%     3.05k ± 0%    ~            (p=0.674 n=10+9)
KVInsert1000_Native-24             23.6k ± 0%     23.6k ± 0%    ~            (p=0.160 n=10+8)
KVInsert1000_SQL-24                27.5k ± 0%     27.5k ± 0%    ~             (p=0.610 n=8+9)
KVInsert10000_Native-24             232k ± 0%      232k ± 0%    ~           (p=0.541 n=10+10)
KVInsert10000_SQL-24                275k ± 0%      275k ± 0%    ~            (p=0.435 n=9+10)
```

Fixes #3299.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="35" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/7605)
<!-- Reviewable:end -->