@nvanbenschoten 2016-07-02
Haven't gotten a chance to look closely yet, but discussed the general approach with @tamird and it seems to all check out. Happy to see this new strategy!