@tschottdorf 2016-07-02
What's with all the assignees? Approach looks good, wonder if there's any shift in benchmark numbers (large inserts, and in particular `InsertDistinct` - and any other benchmark that actually makes nontrivial use of the command queue). I'll leave the detailed review in the capable hands of @nvanbenschoten.