@tschottdorf 2016-07-02
Ok, works for me. In that case, you should call out desired primary reviewer(s) in the PR message.