@tamird 2016-07-02
Added benchmark results from a GCE run. Definitely not what I had hoped to see.