@bdarnell 2016-07-03
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

Regarding assignees, I'd prefer to have 1 assignee (or occasionally 2) per PR (the person who is expected to do a thorough review and give the official LGTM), with other "people who might care" notified with an `@mention` in the PR description. 

---

Review status: 0 of 3 files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[storage/replica.go, line 912 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7605#-KLe5QL4j3SjUt5wBqQv-r2-912:-KLlPUGtB-G4WqIeG32w:-1010767062) ([raw file](https://github.com/cockroachdb/cockroach/blob/ddafcd85fd2cd5af7f6341caa48cffe0141b6dc9/storage/replica.go#L912)):*
<details><summary><i>Previously, tschottdorf (Tobias Schottdorf) wrote…</i></summary>
> I wouldn't add these optimizations unless there's a good reason to do so (I mean letting the caller go early).
</details>
Isn't letting the caller go early the whole reason for the move to channels? If we're going to force the caller to wait we might as well stick with the WaitGroup. 

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/7605#-:-KLlPzyLPGeexaNi6300:-730460689)*
<!-- Sent from Reviewable.io -->