@tschottdorf 2016-07-03
Reviewed 3 of 3 files at r1, 2 of 2 files at r2, 3 of 3 files at r3, 3 of 3 files at r4.
Review status: all files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[storage/replica.go, line 912 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7605#-KLe5QL4j3SjUt5wBqQv-r2-912:-KLlV-0cWVfToKZGX0B3:873293221) ([raw file](https://github.com/cockroachdb/cockroach/blob/ddafcd85fd2cd5af7f6341caa48cffe0141b6dc9/storage/replica.go#L912)):*
<details><summary><i>Previously, bdarnell (Ben Darnell) wrote…</i></summary>
> Isn't letting the caller go early the whole reason for the move to channels? If we're going to force the caller to wait we might as well stick with the WaitGroup. 
</details>
Yes, but we are currently running at one long debug session per innocuous-looking optimization. There's really no good reason to switch to channels here at this point.
With this change, we could deadlock in the command queue and the clients wouldn't block, instead we would leak goroutines over time as their contexts time out. There's no added visibility into the command queue with this change, instead the situation complicates. Hence, my preference for KISS.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/7605)*
<!-- Sent from Reviewable.io -->