@tschottdorf 2016-07-03
Review status: all files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[storage/replica.go, line 912 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7605#-KLe5QL4j3SjUt5wBqQv-r2-912:-KLlVj5UkJK-H1A3vFZx:-972976684) ([raw file](https://github.com/cockroachdb/cockroach/blob/ddafcd85fd2cd5af7f6341caa48cffe0141b6dc9/storage/replica.go#L912)):*
<details><summary><i>Previously, tschottdorf (Tobias Schottdorf) wrote…</i></summary>
> Yes, but we are currently running at one long debug session per innocuous-looking optimization. There's really no good reason to switch to channels here at this point.
> With this change, we could deadlock in the command queue and the clients wouldn't block, instead we would leak goroutines over time as their contexts time out. There's no added visibility into the command queue with this change, instead the situation complicates. Hence, my preference for KISS.
</details>
In case that wasn't clear, I'm feeling strongly about this. I spent all of Friday trying to fix the registration cluster. A lot of the stack traces I saw was things blocked indefinitely on the command queue. While the command queue itself may not have been the problem, I don't want more artificial moving parts here due to an optimization that looks "fun to put in".

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/7605)*
<!-- Sent from Reviewable.io -->