@bdarnell 2016-07-03
Review status: all files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[storage/replica.go, line 912 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7605#-KLe5QL4j3SjUt5wBqQv-r2-912:-KLlZOan7Wf8h7QYkN-c:-2135395974) ([raw file](https://github.com/cockroachdb/cockroach/blob/ddafcd85fd2cd5af7f6341caa48cffe0141b6dc9/storage/replica.go#L912)):*
<details><summary><i>Previously, tschottdorf (Tobias Schottdorf) wrote…</i></summary>
> In case that wasn't clear, I'm feeling strongly about this. I spent all of Friday trying to fix the registration cluster. A lot of the stack traces I saw was things blocked indefinitely on the command queue. While the command queue itself may not have been the problem, I don't want more artificial moving parts here due to an optimization that looks "fun to put in".
</details>
Agreed. My point is that the "optimization" isn't really separable from the rest of the change. If we don't allow clients to return early, we shouldn't make any part of this change. And if we haven't been seeing problems that would be improved by this change it's probably best to leave it alone. 

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/7605)*
<!-- Sent from Reviewable.io -->